food-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai>=1.3.0
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0
httpx>=0.24.0
